Refactor: Move filter functionality into SearchBar component
