+++ src/components/index.js
@@ -1,1 +1,2 @@
 export * from './ListItem';
+export * from './SearchBar';
